RTS: Self-contained heap representation for bignums
To simplify GCs, we store the libtommath bignums as self-contained
objects on the heap (so for the purposes of the GC, they are just data).

The idea is that the `TAG_BIGINT` object stores both the `mp_int` record
and the `mp_digit *` array. See `rts/motoko-rts/src/bigint.rs` for more
details.