Fix #2317: Store mutability in `NewObjE`
The bug happened because the backend expects the `exp.note.Note.typ` of a
`NewObjE` object to be “tight”, and include every field (to find out
whether it is mutable or not). But `Check_ir` doesn’t enforce it, and the
`async/await` translation breaks this (and due to subtyping that is
somewhat ok).

This fixes it by storing the mutability in the AST, and use the type
notes less.