Value.Numerics: Refactoring to prepare word deprecation
this extracts, off #2309, the refactoring of `value`, without any
user-visible changes.

This should make reviewing #2309 easier, and also pave the way to
deprecating the word type.

Because this is based off #2309, it includes some changes to the prims.
In particular, the `num_conv_t1_t2` family is now for trapping
conversions, and `num_warp_t1_2` is for wrapping (currently: to Word,
and between fixed types).