[mo-doc] Generates cross-references for types in signatures (#2278)
# Table of Contents

1.  [Story: Developer can generate links in their docs](#orgc95387b)
    1.  [Design](#orgce62912)
    2.  [Implementation](#org46f7577)
        1.  [The Xref module](#org9be3721)
        2.  [The Namespace module](#orgedf705b)


<a id="orgc95387b"></a>

# Story: Developer can generate links in their docs

Fixes <https://github.com/dfinity/dx-triage/issues/33>

Implements cross-references for types in documented signatures.

To try this out you can build this branch, and then build docs for base:

    cd motoko-base
    /path/to/mo-doc --format html
    firefox docs/index.html

The same works for `--format adoc` but you need a firefox extension to easily preview this:
<https://github.com/asciidoctor/asciidoctor-browser-extension>


<a id="orgce62912"></a>

## Design

The first iteration will just provide links to types defined in the current library. No cross-package referencing yet.
We also don't need to cover *every* edgecase. As long as we don't crash or produce non-sensical output not solving the general problem is fine.
Top-level type definitions or module exports will probably cover over 90% of the types written down in your typical Motoko program.


<a id="org46f7577"></a>

## Implementation


<a id="org9be3721"></a>

### The Xref module

The new `xref` module holds a datatype that lets us describe every entity inside a Motoko module we might want to reference.
It's used as the basis for generating links and anchors in Adoc and Html.
It would also be a good fit for letting users write references in their documentation, if we come up with a textual format for it.


<a id="orgedf705b"></a>

### The Namespace module

The implementation boils down to resolving all `PathT` to their respective definition sites.
Doing this at full fidelity is basically impossible without a recursive traversal passing down an environment.
I went with a reified \`Namespace\` type instead, because I assumed that would be easier to test and work with.

I'm not sure that turned out to be true, but the current implementation is good enough to resolve all paths in `base`,
which is basically what `mo-doc` is used for at this point in time.

It doesn't respect all the shadowing rules of Motoko. At some point I/we could revisit this and try to write it as the recursive traversal instead.