Deprecation as field attribute
this is a variant of #2336. Now, deprecation is an attribute of public
fields of objects (module in particular).

The syntax is `public deprecated "reason"`, so only public fields can be
deprecated. The syntax works for type and value fields alike (big
plus!).

The `type` type is unaffected, so no need to deal with a fake `Depr`
type former, this is nice.

Most of the diff churn is due to `field` now having an extra field that
we usually ignore. Otherwise, this seems nicely orthogonal.

Also todo: Warn when using object patterns to project out a deprecated
field.

Right now, the annotation can only be set on object literals, but there
is no syntax for doing this kind of annotation on a type (e.g. `module
{} : module { foo : deprecated "expl" t }`). Needed?

The `pub_fields` functions can maybe be refactored a
bit (don’t pass through an accumulator, but pass up a map, and add
deprecation annotation), I’ll do that if we go forward with this design.