Remove arithmetic on word types
part of Phase 2 of word deprecation (#2362).

This is removing the operations in a relatively crude, but minimally
invasive way; no need to invest time into different structure in, say,
`operator.mo` when we remove Word types completely a few weeks later.

Should probably be merged early next week (around Monday, Feb 22), so
that users have a chance to see and act on the deprecation warnigs that
we put in the last release.