Deprecate word types
every use of the `WordN` type constructor now causes a warning. Kinda
noisy, but will have the desired effect.

Part of word deprecation phase 2 (#2362)

Should probably be merged early next week (around Monday, Feb 22), so
that users have a chance to see and act on the deprecation warnigs that
we put in the last release.