Prim: Expose raw access to certified data
as per discussions in
https://dfinity.slack.com/archives/CGA566TPV/p1613656127252700
and
https://dfinity.slack.com/archives/CGA566TPV/p1613712967272400
it seems possible that the asset canister (currently written in Motoko)
ought to certify it’s data. While it is unclear if we really want to
implement that in Motoko, the reason not to do it shouldn’t be “the
Motoko team is not exposing this functionality”. So at least expose it
in “raw” form via `mo:prim`.

This could actually be then wrapped in a reasonably usable library.