Put prelude/internals/prim into their own `.mo` files (#2384)
since we are using the `[%blob …]` prepreocessor already in
`lang_utils/error_codes.ml`, there is no reason not to use it also in
`prelude/prelude.ml`, and keep the prelude and prim modules in separate
files. This is mostly for our own editing convenience:

 * We get the right syntax highlighting,
 * line numbers from error messages about prelude or prim line up properly,
 * we can auto-format them once we have the auto-formatter,
 * (maybe) useful for `mo-doc` processing, to have a source file to link to.

While making that change I also moved all the `@`-prefixed definitions
from `prelude.mo` into its own file, `internals.mo`. This comes at the
cost of a few more lines in `pipeline.ml`, but now we have a
`prelude.mo` that really only contains the user-visible prelude.

I checked if maybe we can skip loading `internals.mo` when interpreting,
but it’s needed from `prim.mo`, so no.