Typing: check type declarations are non-expansive (#2387)
Enforces the non-expansiveness criteria identified first by Viroli and adopted by Pierce and Kennedy to ensure termination of sub-typing in *nominal* type systems with generics and subtyping.

Given a set of mutually recursive type definitions, construct a graph whose vertices are the formal parameters (identified by position), `C#i`, and two sorts of labelled edges:

* an occurrence of parameter `C#i` as immediate `j`-th argument to type `D<..,C#i,..>`, adds a (non-expansive) 0-labelled edge `C#i` -0-> `D#j`.

* an occurrence of parameter `C#i` as as a proper sub-term of the `j`-th argument to type `D<...,T[C#i],..>` adds an (expansive) 1-labelled edge `C#i -1-> D#j`.

The graph is expansive iff it contains a cycle with at least one expansive edge.

Non-expansiveness (hopefully) ensures that the set of types encountered during sub-typing, and added to the visited set, is finite. In nominal systems, a visited set is used to reject a cyclic sub-type check, but in our co-inductive setting, it is used to succeed a cyclic check. In either case, it is used terminate the procedure, which is what we care about.

To detect the existence of a cycle, we construct the plain graph obtained by deleting labels (possibly identifying edges), compute its strongly connected components (sccs), and then check whether there is a 1-weighted edge (in the original graph) connecting any two vertices of the *same* component of the  sccs.

Note: even with this checked we still need to enforce that type definitions are contractive to avoid looping in normalization steps.