Smarter typing of binary operators (#2341)
Allow left-to-right or right-to-left type propagation for binary and relational operators.

The decision to switch to checking mode for one side is based on a syntactic criterion for "explicit expressions", a subgrammar of expressions for which inference is guaranteed to deliver "precise" results. If either of the operands is explicit but the other is not, then the explicit operand is inferred first and the other typed in checking mode.

This is sort of a dual (and a canonical generalisation) of the idea to treat literal operands specially.

Fixes #853.