Describe our CI/CD requirements and current implementation (#2403)
in preparation for a meeting with @ali-dfinity, let us try to collect
all the different ways in which we rely on Github/Github
Actions/Hydra/Mergify, document what we want to achieve, and how. This
will be useful documentation even after that meeting (assuming we keep
it up to date).

For this document, the details of our _build system_ (e.g. using `nix`,
`nix-shell`, `make`, etc.  and all the things we do with nix) are out of
scope.