Check_ir: Check for aliasing in IR
the IR nodes have mutable nodes, so aliasing is a bit dangerous, and the
`Const` analysis explicitly relies on that.

So let’s make sure that `Check_ir` complains if there is sharing.

We do so by numbering the `Check_ir` runs, storing that number in the
nodes, and complaining if we see the current numbers.

As a nice side effect, this means that we’d notice if `Check_ir` visits
a node twice.

It seems that we had sharing, but only in literals.