Build users guide on CI
this fixes #1415.

It is very hard to build the documents antora-style, mainly because antora likes to work _on_ repositories, not _within_. So I added code that takes the Antora navigation index, and turns it into a `.adoc` file that includes the others. This way, `make index.html` in `docs/` produces a nice, self-contained file, which is probably good enough if we want to preview our changes.

Hydra will show this at a URL like
https://hydra.dfinity.systems/job/dfinity-ci-build/motoko/docs/latest/download/1/index.html