Vendor wasm-profiler in this repo
the small utility code that is behind our performance test suite flame
graphs was previously at the public repository `dfinity/wasm-profiler`,
until someone noticed that it should have gone through a proper open
sourcing processes before being published there, and it was made a
private repo.

But it’s kinda too small to distract our open sourcing bureaucracy with
deciding whether it can be open sourced, and we _certainly_ don’t want
to get that on the critical path to open sourcing Motoko.

So let’s just declare this part of the “build tooling inside the Motoko
repository” (after all, this is where it is used) and include the code here.

Once we have open sourced Motoko, we can probably split it into a
separate repo again (which makes sense IMHO).