Exploit polarities of type parameters to inform default instantiation (#2448)
Fixes  #1639 (Type Inference: Unexpected implicit instantiation at None)

This extends type parameter inference to, in the absence of an expected return type, use the polarities of the  function's type parameters in the result to choose a principal instantiation, when one exists.
Invariant parameters that are unresolved by inference don't have principal solutions and are rejected as ambiguous.

Note that, though I believe correct, this is actually inconsistent with the way Motoko infers types for (invariant) mutable arrays and var declarations. As pointed out in #1193, Motoko cannot infer infer principal types for `var` and mutable arrays and makes an arbitrary, non principal, choice based on the type of the initializer/lub of the elements.
For consistency, for invariant type parameter instantiation, when under-constrained, we may want to, instead of rejecting, warn and (similarly) use the non-principal inferred upper bound instead. 
I don't really like that idea, but wanted to point out the option.
To see the disparity, with this PR you get the following behaviour:

```
import Prim "mo:prim";

do {
  let a = Prim.Array_init(1, 0); // rejected as ambiguous
};

do {
  let a = Prim.Array_init(1, 0) : [var Int]; // accepted
  a[0] := -1;
};

do { 
  let a = [var 0]; // accepted as [var Nat], but did the programmer actually intend [var Int]?
  a[0] := -1; // rejected
};

do { 
  let a = [var 0] : [var Int ]; // accepted as [var Int] 
  a[0] := -1; // accepted
};
```

- [x] refactor polarity calculation to return map to lattice 
- [x] clean up initial setup
- [x] suppress/adjust polarity info in errors?
- [x] pretty-print types in bi_match.ml
- [x] investigate making l and u maps total, not partial, to simplify code.
- [x] doc comments
- [x] merge claudio/next-moc-polarities and update sources.json to match
- [x] test on parsec (some instantiations needed, but not many more cf. https://github.com/crusso/mo-parsec/pull/2)