Fix Blob.fromArray
I was using `Blob.len` internally, but did not notice that it would
return the length in vanilla representation, i.e. shifted to the left.

This fixes it, by letting `Blob.len` return the raw length, and do the
boxing/shifting to `Nat` only for the implementation of `b.size()`.