Fix moc build errors with OCaml 4.12.0 (#2533)
OCaml 4.12 adds `List.compare`, which breaks this code in compile.ml.

With this PR it builds with both 4.10 and 4.12.

OCaml 4.12 support needed to run https://github.com/ocaml/ocaml-lsp. There is
probably a way to run the latest version of ocaml-lsp with OCaml 4.10, but I
couldn't figure it out, and fixing moc for OCaml 4.12 was easy.