CI: Automatically create tags for releases (#2570)
Whenever we push a tag, this will:

 * Fetch these assets from the cache
   ```
   motoko-linux64-$VERSION.tar.gz
   motoko-macos-$VERSION.tar.gz
   moc-interpreter-$VERSION.js
   moc-$VERSION.js
   ```

 * Extracts the changes from the changelog, failing hard if it doesn’t _precisely_
   look as expected. This also means that nothing happens if you use some tag
   like `some-revision-I-wanted-to-remember`

 * Creates a Github Release, with the body of the changelog entry as the body,
   and the above files as assets. This means one can download motoko from
   ```
   https://github.com/dfinity/motoko/releases/download/0.1.2/motoko-linux64-0.1.2.tar.gz
   ```

Note that the `nix-build` step needs stuff from both linux and darwin. This
will fail if the nix cache was not warm. This could happen if you push the tag
while there were still jobs runnig. In that case, just restart the github
action manually.

With this we can stop deploying via DFINITY’s internal infrastructure.

This also removes the hydra-related infrastructure for releases, and updates `CI.md`.

Fixes #2566