New GC: in-place compaction (#2522)
This merges #2223 and master and enables the new collector with a new moc flag
`--compacting-gc`. The goal is to (1) avoid bitrot in #2223 (2) allow easy
testing and benchmarking of different collectors.

We include both collectors in the Wasms. Binary sizes grow between 0.8% (CanCan
backend) to 1.8% (in simple tests in `run-drun`).

Some benchmark results here:
https://github.com/dfinity/motoko/issues/2033#issuecomment-842379468

An improvement to the compacting GC is currently being implemented in branch
`osa1/compacting_gc_3`.