Compacting GC: get object size before threading fields
This fixes a bug when an object refers to itself. If we thread the
fields first, by the time we get the object's size the header will be
gone. So we need to get the size first, then thread the fields.

Bug caught by #2573 and a test will be added with that PR.