Compacting GC: merge threading and marking
This PR merges threading of object fields with marking. The result is
-14% in instructions in CanCan backend benchmark.

Changes:

- Mark phase of compacting GC now threads fields of marked objects.
  After marking all fields are now threaded.

- Because all fields are threaded during marking, the pass for threading
  is now gone.

- However we still need two passes for unthreading and moving the
  objects (see below). So the first pass after marking unthreads fields.
  Then another pass moves objects.

Original idea was to merge marking and threading, and doing unthreading +
moving in one pass. It turns out that doesn't work. To see why, suppose the
heap is like this:

    | ..., object N, ..., object M, ... |

where N and M point to each other. If we unthread + move (slide) in one pass,
then after sliding N, it's possible that live objects between N and M will be
slid to N's original location. So when we unthread M which will have a pointer
to N's old location in its header we'll read random data from whatever's moved
to N's original location.

The solution is to do these in different passes: first unthread, then slide.

The result is still 3 passes as before, but for some reason that I don't
understand, this version is 14% faster on CanCan backend benchmark (42,005,245
instructions to 35,775,591).

I think "An Efficient Garbage Compaction Algorithm" is describing a way to
implement in-sliding compaction in two passes instead of three, but I found
that paper difficult to follow. I will revisit it now to see if this can be
improved further.