ci-pr.nix: Fix wasm-hash-for
in 58219bd06074ac924fe63d2007fa4bcf7106d9a3 I broke this script: The way
I used timeout was broken. This meant that we never produce any `.wasm`
files, which meant that the pr diff never noticed a differnce, and thus
didn’t print any performance regressions.

This fixes the bug and adds a safeguard that fails if no wasm is
generated.

Note that this compiles all files with `-ref-system-api`, meaning that
many tests in `run-drun` are not taken into account.