Smarter GC scheduling
We now avoid doing GC after every message and only do when these hold:

1. Heap grows more than HEAP_GROWING_FACTOR since last GC
2. More than SMALL_HEAP_DELTA is allocated
3. Current heap size is less than MAX_HP_FOR_GC

(1) is to make sure we don't do GC when heap size does not change much.
On small heaps `heap_size * HEAP_GROWING_FACTOR` will be a small amount,
so (2) is to avoid doing redundant GCs in such cases. (3) is to make
sure we'll have as allocation are as possible for the next message when
heap is larger than some amount (MAX_HEAP_FOR_GC).

A few notes/todos:

- Tests in run-drun do not allocate much, so this scheduling effectively
  disables GC completely for those tests. We should probably do GC after
  every message in tests as before, otherwise test coverage for the
  collectors will be reduced significantly.

- Scheduling is currently not tested in RTS test suite.