Compacting GC: thread backwards pointers in marking (#2647)
This PR implements the mark-compact GC algorithm briefly described by
"High-Performance Garbage Collection for Memory-Constrained Environments"
section 5.1.2 (I think the original paper is "An Efficient Garbage Compaction
Algorithm" but that paper is a bit difficult to follow).

The idea is as follows: when marking an object we thread backwards pointers of
that object. After we linearly scan the heap as before. For a live object, we
unthread it and update backwards pointers to it to the objects new location,
move the object, and then thread its forwards pointers. After the pass all
objects are moved and all references are updated.

This reduces number of passes in the original mark-compact collector to 2.

CanCan backend benchmark reports -19.8% in instructions. Compared to copying
GC, mark-compact GC is now 36% slower, instead of 70% as before.

This PR also optimizes `mark_static_roots` by inlining `mark_fields` in the
body and specializing it for `MutBox`: static root array only points to static
`MutBox`es so no need to call more general `mark_fields`. This optimization
gives us approximately -2% in instructions.

See also #2641 for another, slower variant of the algorithm.