RTS: Introduce heap objects for unused space
Two new objects OneWordFiller and FreeSpace introduced to mark unused
spaces in heap.

These are mainly needed in #2706 and for new gc experiments, but they
are also useful to mark the slop when shrinking blobs in principal id
functions.

Previously when shrinking a blob we were filling the unused space with
zeros, and the gc was skipping the zeros when scanning the heap for e.g.
debug prints or sanity checks.

With OneWordFiller and FreeSpace, for 1 and 2 words unused space we do
the same amount of writes (header or header + length), for anything
larger we do 2 writes, instead of one write per word.