Nix gardening (#2670)
This  is in preparation for the `stdenv.lib` removal. Strangely, the warning persists. I assume it comes from some expressions that are outside of this repo and we only transitively pull in.

This is the warning we eventually want to fix:
```
trace: Warning: `stdenv.lib` is deprecated and will be removed in the next release. Please use `lib` instead. For more information see https://github.com/NixOS/nixpkgs/issues/108938
```