Use deterministic hashers in RTS test suite
By default Rust's HashMap and HashSet hashers are non-deterministic,
which is a problem when debugging as the tests can fail differently in
consecutive runs. I even observed that a buggy test passes when I run
the test suite repeatedly a few times.

This PR switches to `fxhash`s HashMap and HashSet, which is just std map
and set with a deterministic hasher. In the future we should be careful
to not use std maps and sets and always use fxhash.

`maplit` is removed as the `hashmap` and `hashset` macros only allow
generating std maps and sets with non-det hashsers.