Fix handling of self references in mark-compact GC, imrpove tests (#2721)
One line fix to thread self references during marking in mark-compact GC.

We had a test with self references but it accidentally passed. The problem was
none of the objects below the object with self reference were dead, so the
object was not moved, and as a result the field with the self reference had the
correct value even though we didn't update it.

The test is updated to add a dead object before the object with self ref. It
fails without the fix.

Test code is also updated to use a vector for the heap description, instead of
a map. This allows specifying object order in heap. With the previous HashMap
based implementation there is no way to reason about the object orders, so for
example the test with a backward pointer might or might not have a backward
pointer, depending on the HashMap (and hasher) implementation. With the vector
representation allocation is done in the same order as the test description,
allowing testing backward and forward pointers.

Thanks to @ulan for the catching the bug.