Implement `Int` ⇔ `Float` conversion as primitives (#2733)
See https://forum.dfinity.org/t/motokos-type-system-converting-between-types.

The language-level change in this PR is that we add two new functions to `prim.mo`:
- `floatToInt : Float -> Int`
- `intToFloat : Int -> Float`

`floatToInt` cannot fail, at least in theory. In the interpreter I have seen failures for `floatToInt(1.7e308)`, while the compiler works well. Tests go up to `1e100`. And then there are `inf` and `NaN`, which are not even tested.

If `intToFloat` overflows, the result is `(+/-)inf`.

Of course these conversions are lossy in general. Caveat emptor!

### Base library implications

Let's merge this PR after soaking a few days/weeks.

Then module `Float.fromInt` would call the new primitives instead of going the `Int` -> `Int64` -> `Float` route (which does trap for big numbers). The fix need to go into the `motoko-base` lib at a later stage. This is the place:
``` swift
  /// Conversion from Int via Int64. May trap.
  public let fromInt : Int -> Float =
    func (x : Int) : Float = fromInt64(Prim.intToInt64(x));
```