CI.md: Mention branch protection
I have changed the branch protection configuration to require the Github
Actions to succeed, and no longer block on hydra. This PR merely updates
our documentation to reflect this.

This means we can (and should) ask the IDX team (if it still has that
name) to disable hydra for the Motoko repo.