nix: Try different patch to fix -pie issues
in #2532 we added a patch related to static building of ocaml packages,
submitted to nixpkgs as https://github.com/NixOS/nixpkgs/pull/124498,
but it was never merged upstream.

Supposedly a patch from https://github.com/NixOS/nixpkgs/pull/135619
fixes it as well (and maybe more properly). So let’s try that!