Prim.trap: Expose ic0.trap
the System API provides canisters with a way to explicitly trap with an
error message. This is very useful for developing, e.g. to get useful
messages in code you think is unreachable, or when checking invariants.

This exposes this as `Prim.trap`. Later this can be exposed in
`motoko-base` via `Debug.trap` maybe, and also the odd `Prelude.nyi`,
`Prelude.xxx` and `Prelude.unreachable` can be implemented that way in a
more helpful way.

Semantically this is like the existing `assert false`, just with better
control over the error message and a more helpful type (return type
`None`).