Set the lowest bit in object tags
This is in preparation for #2790 and #2706. With #2790 we will start
using rest of the headers for GC metadata (set some of the high bits)
which will break compacting GC as we won't be able to distinguish a
heap address from an object header by checking if the value is larger
than the max. tag value. This check assumes a heap address cannot be
smaller than the max. tag value, which holds because we have at least 64
KiB Rust stack, and then static data for the canister.

With the high bits of headers set, it's possible that some of the
headers will have a value larger than 64 * 1024. So the current check no
longer works.

To allow distinguishing heap locations from headers, this PR refactors
objects tags so that they will all have the least significant bit set.
Since objects and fields are all word aligned (so have the lowest 2 bits
unset, this invariant was established in #2764), we can now check the lowest
bit and distinguish an address from a header.