introduce `NextArrayOffset` primitive
This still operates on `Nat` values (will change later)
but assumes a restricted numeric range of operands and
as such can be performed without resorting to proper bignum
arithmetic calls. It is basically incrementing by
one now.

The master plan is that the interpreter will do
`Nat` arithmetic, but the compiler will work
with byte offsets. But we need a few more primitives
for that to work.