Candid decoding: Skip `vec any` fast
this mitigates a possible candid decoding cycle bomb: A large `vec any`
or `vec reserved` vector in a position where it is ignored by subtyping.
Such a vector, even of length 2^31, is very compact in the encoding (a
few bytes), but a naive decoder will still run 2^31 iterations of
skipping nothing.

We can easily detect this by observing that `skip_any` does not move the
pointer.

Does not fix cycle (and memory) bombs where the type of the method
actually contains a type that is a supertype of `vec null` (e.g. `[?…]`)