Refactor RTS nix derivation `allDeps`: (#2877)
When the copied directories contain vendored dirs of same package, the
current code fails with errors like

    cp: cannot create regular file '/nix/store/nmbhnj7hmx6piqh3nylwjbipkqnk2ssm-merged-rust-deps/cfg-if-0.1.10/.cargo-checksum.json': Permission denied

This currently happens in #2706.

One way to fix this is by making the directory writable, with `chmod -R
+w $out` after the first `cp`.

It turns out there's a `nixpkgs` function for doing merging contents of
two directories: `symlnkJoin`. It's more concise and it solves the
problem.

Thanks @basvandijk for the tip.