Backport random heap tests from osa1/page_alloc (#2889)
This implements a very simple random heap generation that caught a lot of bugs
in #2706. The idea is similar to QuickCheck/proptest style property testing: we
generate random heaps using a pseudo-random number generator and a fixed seed,
and then run the GC tests as usual (run both collectors, check GC properties).

Difference from QuickCheck/proptest-style testing is we don't implement
shrinking. Implementing shrinking manually is not a lot of work, but I believe
it won't work in practice, because checking subsets of the nodes/edges of the
graph will take exponential time (N nodes will have 2^N subsets of nodes), and
we really need to check all subsets as these graphs won't have any properties
that allows us to shrink more efficiently.